pyproject.toml


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cbfs"
version = "0.1.0"  # Update version for new releases
description = "Floating Feature Selection guided by Strangeness Quantification "
authors = [
    { name = "Marcos López-De-Castro", email = "mlopezdecas@unav.es" }
]
readme = "docs/README.md"
license = { text = "MIT" }
requires-python = ">=3.10.5"
dependencies = [
    "scikit-learn>=1.7.1,<2",
    "numba>=0.61.2,<0.62",
    "joblib>=1.5.2,<2",
    "threadpoolctl>=3.6.0,<4",
    "pyreadr>=0.5.3,<0.6",
    "scanpy>=1.11.5,<2",
    "igraph>=0.10.0,<0.12",
    "python-igraph>=0.11,<0.12",
    "damply>=0.9.0,<0.11",
    "leidenalg>=0.10",
    "neurocombat-sklearn>=0.1.3,<0.2"
]



[tool.hatch.build.targets.wheel]
packages = ["src/cbfs"]

[tool.hatch.build.targets.sdist]
include = ["src/**", "pyproject.toml", "docs/README.md", "LICENSE"]